onebinary: move up to mistserver.cpp
