+++ src/output/mist_out.cpp
@@ -1,5 +1,3 @@
-#include "output_rtmp.h"
-#include "output_hls.h"
 #include <mist/config.h>
 #include <mist/defines.h>
 #include <mist/socket.h>
@@ -27,7 +25,7 @@ void handleUSR1(int signum, siginfo_t *sigInfo, void *ignore){
 }
 
 template<class T>
-int Main(int argc, char *argv[]){
+int OutputMain(int argc, char *argv[]){
   DTSC::trackValidMask = TRACK_VALID_EXT_HUMAN;
   Util::redirectLogsIfNeeded();
   Util::Config conf(argv[0]);
@@ -74,28 +72,5 @@ int Main(int argc, char *argv[]){
     }
   }
   INFO_MSG("Exit reason: %s", Util::exitReason);
-  return 0;
-}
-
-int main(int argc, char *argv[]){
-  if (argc < 2) {
-    INFO_MSG("usage: %s [MistSomething]", argv[0]);
-    return 1;
-  }
-  // Create a new argv array without argv[1]
-  int new_argc = argc - 1;
-  char** new_argv = new char*[new_argc];
-  for (int i = 0, j = 0; i < argc; ++i) {
-      if (i != 1) {
-          new_argv[j++] = argv[i];
-      }
-  }
-  if (strcmp(argv[1], "MistOutHLS") == 0) {
-    return Main<Mist::OutHLS>(new_argc, new_argv);
-  }
-  else if (strcmp(argv[1], "MistOutRTMP") == 0) {
-    return Main<Mist::OutRTMP>(new_argc, new_argv);
-  }
-  INFO_MSG("binary not found: %s", argv[1]);
   return 0;
 }
